Misc: Moving items here and there for better alignment with folder definitions, that is utils should only have helper items not actual code used for running things
