+++ resources/assets/README.md
@@ -1,0 +1,3 @@
+# Assets
+
+This folder has all map assets that you can use
