+++ resources/examples/README.md
@@ -1,0 +1,3 @@
+# Examples
+
+Will use this for storing examples
